requirements.txt
apipkg==1.5
appdirs==1.4.4
argcomplete==1.12.2
asgiref==3.3.1
asttokens==2.0.4
attrs==20.3.0
base58==2.1.0
bitarray==1.2.2
black==19.10b0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cytoolz==0.11.0
distro==1.5.0
Django==3.1.7
djangorestframework==3.12.2
eth-abi==2.1.1
eth-account==0.5.2
eth-brownie==1.13.1
eth-event==1.2.1
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
execnet==1.8.0
hexbytes==0.2.1
hypothesis==5.41.3
idna==2.10
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.7.0a1
jsonschema==3.2.0
lru-dict==1.1.7
more-itertools==8.7.0
multiaddr==0.0.9
mythx-models==1.9.1
netaddr==0.8.0
packaging==20.9
parsimonious==0.8.1
pathspec==0.8.1
pipx==0.16.0.0
pluggy==0.13.1
prompt-toolkit==3.0.8
protobuf==3.15.0
psutil==5.8.0
py==1.10.0
py-solc-ast==1.2.8
py-solc-x==1.1.0
pycodestyle==2.6.0
pycryptodome==3.10.1
Pygments==2.6.1
pygments-lexer-solidity==0.5.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.0.1
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
pythx==1.6.1
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
rlp==1.2.0
semantic-version==2.8.5
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
toml==0.10.2
toolz==0.11.1
tqdm==4.53.0
typed-ast==1.4.2
urllib3==1.26.3
userpath==1.4.2
varint==1.0.2
vvm==0.1.0
vyper==0.2.10
wcwidth==0.2.5
web3==5.11.1
websockets==8.1
